RHDC- Result Calculator - Adding new Algorithm to calculate results by RPR and TS including addition of Algorithm settings entity

+++ Core/Models/Algorithm/AlgorithmResult.cs
@@ -7,7 +7,7 @@ namespace Core.Models.Algorithm
     public class AlgorithmResult
     {
         public int AlgorithmId { get; set; }
-        public decimal Accuracy { get; set; }
+        public double Accuracy { get; set; }
         public int RacesFiltered { get; set; }
 
     }
